requirement.txt
PyMySQL
python-slugify
requests
beautifulsoup4
pdfminer.six
